# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minhas Leituras Documentação",
        "description": "Gerenciamento de Leituras Este projeto é uma aplicação de gerenciamento de leituras onde os usuários podem cadastrar, organizar e acompanhar seu progresso de leitura de livros, HQs e outros materiais. O sistema permite que os usuários gerenciem suas leituras, definam metas, escrevam avaliações, recebam lembretes para ler, e muito mais.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/usuario-leitura/{id_usuario_leitura}": {
            "delete": {
                "tags": [
                    "Usuario Leituras"
                ],
                "summary": "Excluir Leitura do Usuário",
                "description": "Excluir Leitura do Usuário pelo ID",
                "operationId": "f2b3d0574321b7a34d7e10221bb83f80",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token de autenticação do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id_usuario_leitura",
                        "in": "path",
                        "description": "ID da leitura do usuário a ser excluída",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leitura excluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Dados adicionais, se houver",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado, token inválido ou ausente"
                    },
                    "404": {
                        "description": "Leitura não encontrada"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/usuario/login": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Realizar Login do Usuário",
                "description": "Endpoint para Login do Usuário",
                "operationId": "fe1bc6703549926f5439b02d016b80c6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "email": {
                                        "description": "E-mail do Usuário",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Senha do Usuário",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação nos campos enviados"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                }
            }
        },
        "/api/usuario/logout": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Realizar Logout do Usuário",
                "description": "Endpoint para logout do usuário autenticado",
                "operationId": "3700977f51662c6a3e5c3cb1cc477986",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Dados adicionais, se houver",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado, token inválido ou ausente"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}